+++ CHANGELOG.md
@@ -2,6 +2,8 @@ UNRELEASED
 ----------
 
 * Correct license in `package.json`;
+* Correct tslint disable scope;
+* Implement subword deletion commands;
 
 0.1.0 2018-07-26
 ----------------